+++ tests/integration/backpropnumkey/db.php
@@ -1,0 +1,18 @@
+<?php
+namespace Wikimedia\Rdbms;
+
+interface IDatabase {
+	public function select(
+		$table, $vars, $conds = '', $fname = __METHOD__,
+		$options = [], $join_conds = []
+	);
+}
+
+class Database implements IDatabase {
+	public function select(
+		$table, $vars, $conds = '', $fname = __METHOD__,
+		$options = [], $join_conds = []
+	) {
+		return 'placeholder';
+	}
+}